Fix compile warnings.

+++ src/lib/threads.c
@@ -49,7 +49,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 // On some systems pthread_t is an int and on others it's a pointer
 static unsigned long long int __rrr_pthread_t_to_llu (pthread_t t) {
-	return t;
+	return (unsigned long long int) t;
 }
 
 #define RRR_PTHREAD_T_TO_LLU(t) \
